chore: doc cleanup and misc changes

+++ pluto_utils/sim_setup.sh
@@ -1,28 +1,58 @@
 #!/bin/bash
 script_dir="$PLUTONLIB/pluto_utils"
 cluster_run=false
+arg_save_dir=""
+
+show_help() {
+    echo "-h = show help"
+    echo "-c = cluster, submits job to cluster to run pluto"
+    echo "-s = arg_save_dir, overwrites save_dir to desired location"
+}
 
 #script options, needs help
-while getopts "c" opt; do
-    case $opt in 
-        c) cluster_run=true ;;
-        *) echo "Usage: $0 [-c]" ;; #[-n cores] [-i ini_file] [-r run_dir]
+while getopts "hcs:" opt; do
+    case $opt in
+        h) show_help
+            exit 0
+            ;; 
+        c) cluster_run=true 
+            ;;
+        s) arg_save_dir=$OPTARG
+            ;;
+        *) echo "Usage: $0" 
+            show_help
+            exit 1 
+            ;;
     esac
 done
 
-echo "Moving to $SIM_DIR" #env_var SIM_DIR
-printf "Current Structure: " && ls "$SIM_DIR"
+if [[ "$arg_save_dir" != "" ]]; then #using opt arg for save dir
+    printf "\n"
+    echo "Setting simulation save directory to $arg_save_dir" 
+    save_dir="$arg_save_dir"
 
-printf "\n"
-read -e -p "Enter Current Simulation Name: " cur_sim
-save_dir="${SIM_DIR}/${cur_sim}"
-mkdir -p "$save_dir" # e.g Make sure the /Jet_xx folder exists
+elif [[ "$arg_save_dir" == "" ]]; then
+    echo "Moving to $SIM_DIR" #env_var SIM_DIR
+    printf "Current Structure: " && ls "$SIM_DIR"
+    printf "\n"
+    read -e -p "Enter Current Simulation Name: " cur_sim
+    save_dir="${SIM_DIR}/${cur_sim}"
+fi
+
+if [ ! -d "$save_dir" ]; then #option to create dir if required or exit if mistake, needs while loop
+    read -p "Create directory? $save_dir [y/n]: " create_dir
+    if [[ "$create_dir" == "y" ]]; then
+        mkdir -p "$save_dir" # e.g Make sure the /Jet_xx folder exists
+    else
+        echo "exiting..."
+        exit 0
+    fi
+fi
 
 echo "Copying .sh/.ini files..." 
 cp "$script_dir/job_submit.sh" "$save_dir/job_submit.sh"
 cp "$script_dir/pluto_template.ini" "$save_dir/pluto_template.ini"
 cp "$script_dir/pluto_run.sh" "$save_dir/pluto_run.sh"
-# cp "$script_dir/clean.sh" "$save_dir/clean.sh"
 cp "$script_dir/sim_setup.sh" "$save_dir/sim_setup.sh"
 
 printf "\n"
